% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirnaGS.R
\name{rbiomirGS_logistic}
\alias{rbiomirGS_logistic}
\title{rbiomirGS_logistic}
\usage{
rbiomirGS_logistic(objTitle = "mirna_mrna", mirna_DE,
  var_mirnaName = "miRNA", var_mirnaFC = "logFC", var_mirnaP = "p.value",
  mrnalist, mrna_Weight = NULL, gs_file = NULL, optim_method = "L-BFGS-B",
  p.adj = "fdr", ..., parallelComputing = FALSE, clusterType = "PSOCK")
}
\arguments{
\item{objTitle}{Output \copde{csv} file name prefix.}

\item{mirna_DE}{DE list of miRNAs of interest. This is a \code{dataframe} object for now.}

\item{var_mirnaName}{Variable name for miRNA names in the DE list. Default is \code{"miRNA"}.}

\item{var_mirnaFC}{Variable name for log transformed miRNA fold change (logFC) in the DE list. Default is \code{"logFC"}.}

\item{var_mirnaP}{Variable name for miRNA p value in the DE list. Default is \code{"p.value"}.}

\item{mrnalist}{List containing the mRNA targets for the miRNAs of interest. This is a \code{list} object and can be obtained from \code{\link{rbiomirGS_mrnalist}} function.}

\item{mrna_Weight}{A vector weight for the miRNA-mRNA interaction. Default is \code{NULL}.}

\item{gs_file}{Input \code{gmt} for gene set, and can be obtained from \code{ensembl} databases.}

\item{optim_method}{The parameter optimization method for the logistic regression model. Options are \code{"L-BFGS-B"}, \code{"BFGS"}, \code{"CG"}, \code{"SANN"}, and \code{"Brent"}. Default is \code{"L-BFGS-B"}.}

\item{p.adj}{P value adjustment methods. Options are \code{"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none""}. Default is \code{"fdr"}.}

\item{...}{Addtional arguments for \code{optim} function.}

\item{parallelComputing}{If to use parallel computing or not. Default is \code{FALSE}.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
A \code{csv} containing GS enrichment analysis results.
}
\description{
Logistic regression-based gene set analysis using measured miRNA p value and fold change,  with the option of custom-setting parameter optimization algorithms.
}
\examples{
\dontrun{
geneset <- rbiomirGS_gmt(file = "kegg.gmt")
}
}
