% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRGSA.R
\name{rbiomirGS_mrnalist}
\alias{rbiomirGS_mrnalist}
\title{rbiomirGS_mrnalist}
\usage{
rbiomirGS_mrnalist(mir = NULL, sp = "hsa", queryType = NULL,
  predictPercentage = 5,
  url = "http://multimir.ucdenver.edu/cgi-bin/multimir.pl",
  parallelComputing = FALSE, clusterType = "PSOCK")
}
\arguments{
\item{mir}{Input miRNAs vector.}

\item{sp}{Species. Options are \code{"hsa"} (default), \code{"mmu"} and \code{"rno"}.}

\item{queryType}{Type of reuslts. Options are \code{"validated"} and \code{"predicted"}.}

\item{url}{The database host server: "http://multimir.ucdenver.edu/cgi-bin/multimir.pl"}

\item{parallelComputing}{If to use parallel computing or not. Default is \code{FALSE}.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{predictPrecentage}{Set only if \code{queryType = "predicted"}. The percentage of the top scored mRNA predictions to return. Default is \code{5}.}
}
\value{
The function returns a \code{list} object with mRNA targets for the query miRNAs.
}
\description{
Obtain target mRNA list for miRNAs of interest. Resuls can be either predicted or validated. The function uses multiple databsaes hosted at \code{multimir.ucdenver.edu/}.
}
\examples{
\dontrun{
mRNA <- rbiomirGS_mrnalist(mir = c("hsa-miR-26a-5p", "hsa-miR-100-5p"), sp = "hsa", queryType = "predicted", predictPercentage = 10)
}
}
