% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirnascan.R
\name{rbiomirgs_mrnascan}
\alias{rbiomirgs_mrnascan}
\title{rbiomirgs_mrnascan}
\usage{
rbiomirgs_mrnascan(objTitle = "miRNA", mir = NULL, sp = "hsa",
  addhsaEntrez = FALSE, queryType = NULL, predictPercentage = 10,
  url = "http://multimir.org/cgi-bin/multimir_univ.pl",
  parallelComputing = FALSE, clusterType = "PSOCK")
}
\arguments{
\item{mir}{Input miRNAs vector. Can't be a matrix or data frame.}

\item{sp}{Species. Options are \code{"hsa"} (default), \code{"mmu"} and \code{"rno"}.}

\item{addhsaEntrez}{When \code{sp = "mmu"} or \code{sp = "rno"}, users can set this argument to \code{TRUE} so that a new list containing hsa ortholog entrez ID will be exported to the environment. The function connects to the up-to-date \code{ensembl} databases.}

\item{queryType}{Type of reuslts. Options are \code{"validated"} and \code{"predicted"}.}

\item{url}{The database host server: \code{"http://multimir.org/cgi-bin/multimir_univ.pl"}}

\item{parallelComputing}{If to use parallel computing or not. Default is \code{FALSE}.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{predictPrecentage}{Set only if \code{queryType = "predicted"}. The percentage of the top scored mRNA predictions to return. Default is \code{5}.}
}
\value{
The function returns a \code{list} object with mRNA targets for the query miRNAs, as well as detailed resutls to the working directory as \code{csv} files. When \code{sp = "mmu"} or \code{sp = "rno"}, and \code{addhsaEntrez = TRUE}, a new list containing hsa ortholog entrez ID will be exported to the environment.
}
\description{
Obtain target mRNA list for miRNAs of interest. Results can either be predicted or validated miRNA-mRNA interactions. The function uses multiple databsaes hosted at \code{multimir.ucdenver.edu/}.This function needs a internet connection.
}
\examples{
\dontrun{
rbiomirgs_mrnascan(objTitle = "mmu_mirna", mir = c("mmu-miR-26a-5p", "mmu-miR-100-5p"), sp = "mmu", addhsaEntrez = TRUE, queryType = "predicted", predictPercentage = 10)
}
}
