% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rbiomirgs_volcano}
\alias{rbiomirgs_volcano}
\title{rbiomirgs_volcano}
\usage{
rbiomirgs_volcano(
  gsadfm,
  topgsLabel = FALSE,
  n = 5,
  gsVar = "GS",
  padding = 0.5,
  gsLabelSize = 5,
  logoddsratio = 0,
  fdr = TRUE,
  q_value = 0.05,
  Title = NULL,
  xLabel = "Log odds ratio",
  yLabel = "-log10(p value)",
  p_line_offset = 1,
  symbolSize = 2,
  sigColour = "red",
  nonsigColour = "gray",
  xTxtSize = 10,
  yTxtSize = 10,
  plotWidth = 170,
  plotHeight = 150,
  boxplot = TRUE
)
}
\arguments{
\item{gsadfm}{Input dataframe. The dataframe is the output from \code{\link{rbiomirgs_logistic}}.}

\item{topgsLabel}{If to display the top GS identification on the plot. Default is \code{FALSE}.}

\item{n}{Number of top GS name to display. Only set this argument when \code{topgsLabel = TRUE}. Default is \code{5}.}

\item{gsVar}{The name of the variable for GS name. Only set this argument when \code{geneName = TRUE}. Default is \code{GS}.}

\item{padding}{When \code{topgsLabel = TRUE}, to set the distance between the dot and the gene symbol. Default is \code{0.5}.}

\item{gsLabelSize}{When \code{topgsLabel = TRUE}, to set the label size. Default is \code{5}.}

\item{logoddsratio}{Threshold for logoddsratio for volcano plot. Default is \code{0}.}

\item{fdr}{If to use fdr corrected p value for plotting. Default is \code{TRUE}.}

\item{q_value}{Threshold for the p value. Default is \code{0.05}.}

\item{Title}{Figure title. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{xLabel}{X-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{yLabel}{Y-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{"Mean Decrease in Accurac"}}

\item{p_line_offset}{Percentage offset for the horizontal line. Default is \code{1} (no off set).}

\item{symbolSize}{Size of the symbol. Default is \code{2}.}

\item{sigColour}{Colour of the significant genes or probes. Default is \code{"red"}.}

\item{nonsigColour}{Colour of the non-significant genes or probes. Default is \code{"gray"}.}

\item{xTxtSize}{Font size for the x-axis text. Default is \code{10}.}

\item{yTxtSize}{Font size for the y-axis text. Default is \code{10}.}

\item{plotWidth}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{plotHeight}{The height of the figure for the final output figure file. Default is \code{150}.}
}
\value{
A \code{pdf} file of the volcano plot.
}
\description{
volcano plot function
}
\details{
The function will use uncorrected p value for thresholding if no significant results were found under fdr. A message will display.

         \code{p_line_offset} is useful when \code{fdr = TRUE} and only one feature is passing the threshold.
         With offset, the horizontal line would no longer pass through the dot.
}
\examples{
\dontrun{

rbiomirgs_volcano(gsadfm = mirna_mrna_gs, n = 5)

}
}
