% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirnaGS.R
\name{rbiomirGS_gmt}
\alias{rbiomirGS_gmt}
\title{rbiomirGS_gmt}
\usage{
rbiomirGS_gmt(file)
}
\arguments{
\item{file}{Input \code{gmt} file.}
}
\value{
The function loads the \code{gmt} gene set database file and returns a \code{list} object.
}
\description{
load \code{gmt} gene set files as lists
}
\examples{
\dontrun{
geneset <- rbiomirGS_gmt(file = "kegg.gmt")
}
}
