% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirnaGS.R
\name{rbiomirgs_logistic}
\alias{rbiomirgs_logistic}
\title{rbiomirgs_logistic}
\usage{
rbiomirgs_logistic(
  objTitle = "mirna_mrna",
  mirnascoreTitle = "mirnascore",
  mrnascoreTitle = "mrnascore",
  defile = NULL,
  mirna_DE = NULL,
  var_mirnaName = "miRNA",
  var_mirnaFC = "logFC",
  ratioFC = FALSE,
  var_mirnaP = "p.value",
  mrnalist = NULL,
  mrna_Weight = NULL,
  gs_file = NULL,
  optim_method = "IWLS",
  p.adj = "fdr",
  ...,
  parallelComputing = FALSE,
  clusterType = "PSOCK",
  verbose = TRUE
)
}
\arguments{
\item{objTitle}{Output GS result \code{csv} file name prefix. Default is \code{mirna_mrna}.}

\item{mirnascoreTitle}{Output miRNA score \code{csv} file name. Default is \code{mirnascore}.}

\item{mrnascoreTitle}{Output mRNA score \code{csv} file name. Default is \code{mrnascore}.}

\item{defile}{The input \code{csv} file containing miRNA list and DE resutls.}

\item{mirna_DE}{DE list of miRNAs of interest. This can be a \code{data.frame}, \code{matrix} or \code{list} object.}

\item{var_mirnaName}{Variable name for miRNA names in the DE list. Default is \code{"miRNA"}.}

\item{var_mirnaFC}{Variable name for miRNA fold change (or log transformed FC) in the DE list. Default is \code{"logFC"}.}

\item{ratioFC}{Whether the FC provided is a ratio value. Default is \code{FALSE}.}

\item{var_mirnaP}{Variable name for miRNA p value in the DE list. Default is \code{"p.value"}. Note that the value will be -log10 transformed before calculating the miRNA score.}

\item{mrnalist}{List containing the mRNA targets for the miRNAs of interest. This is a \code{list} object and can be obtained from \code{\link{rbiomirgs_mrnascan}} function.}

\item{mrna_Weight}{A vector weight for the miRNA-mRNA interaction. Default is \code{NULL}.}

\item{gs_file}{Input \code{gmt} for gene set, and can be obtained from \code{ensembl} databases.}

\item{optim_method}{The parameter optimization method for the logistic regression model. Options are \code{"L-BFGS-B"}, \code{"BFGS"}, and \code{"IWLS"}. Default is \code{"IWLS"}.}

\item{p.adj}{P value adjustment methods. Options are \code{"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}. Default is \code{"fdr"}.}

\item{...}{Additional arguments for \code{optim} function.}

\item{parallelComputing}{If to use parallel computing or not. Default is \code{FALSE}.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{verbose}{Whether to display}
}
\value{
A \code{data.frame} object containing GS results to to the environment, as well as a \code{csv} to the working directory. A \code{txt} file containing the iteration information will be generated if \code{opti_method = "L-BFGS-B" or "BFGS"}.
}
\description{
Logistic regression-based gene set analysis using measured miRNA p value and fold change,  with the option of custom-setting parameter optimization algorithms.
}
\examples{
\dontrun{
rbiomirgs_logistic(objTitle = "mirna_mrna",
                   mirna_DE = tstdfm2, var_mirnaName = "miRNA", var_mirnaFC = "logFC", var_mirnaP = "pvalue",
                   mrnalist = hsa_mrna_entrez_list_woNA, mrna_Weight = NULL,
                   gs_file = "~/OneDrive/my papers/my papers/potential_DRDC_paper 2 (diving)/data/kegg.v5.2.entrez.gmt",
                   optim_method = "L-BFGS-B", p.adj = "fdr",
                   parallelComputing = FALSE, clusterType = "PSOCK")
}
}
